aws: Rearrange private functions
Trying to maintain the function sequence of clusterinfo.go
