+++ pkg/cloudclient/aws/aws.go
@@ -22,7 +22,6 @@ import (
 	configv1 "github.com/openshift/api/config/v1"
 	cloudingressv1alpha1 "github.com/openshift/cloud-ingress-operator/pkg/apis/cloudingress/v1alpha1"
 	"github.com/openshift/cloud-ingress-operator/pkg/config"
-	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -135,18 +134,6 @@ func readClusterRegionFromConfigMap(kclient client.Client) (string, error) {
 		return "", err
 	}
 	return parseClusterRegionFromConfigMap(cm)
-}
-
-//getClusterRegion returns the installed cluster's AWS region
-func getClusterRegion(kclient client.Client) (string, error) {
-	infra, err := baseutils.GetInfrastructureObject(kclient)
-	if err != nil {
-		return "", err
-	} else if infra.Status.PlatformStatus == nil {
-		// Try the deprecated configmap. See https://bugzilla.redhat.com/show_bug.cgi?id=1814332
-		return readClusterRegionFromConfigMap(kclient)
-	}
-	return infra.Status.PlatformStatus.AWS.Region, nil
 }
 
 // NewClient creates a new CloudClient for use with AWS.
